requirements.txt
discord.py
requests==2.18.4
websockets==6.0
aiohttp==3.5.4
attrs>=17.3.0
async-timeout<4.0,>=3.0
yarl<2.0,>=1.0
multidict<5.0,>=4.0
chardet<4.0,>=2.0
idna>=2.0
psycopg2